+++ pkg/httputils/httputilsimplementationindependend/GenericResponse.go
@@ -1,8 +1,8 @@
 package httputilsimplementationindependend
 
 import (
-	"github.com/asciich/asciichgolangpublic/fileformats/jsonutils"
-	"github.com/asciich/asciichgolangpublic/fileformats/yamlutils"
+	"github.com/asciich/asciichgolangpublic/pkg/fileformats/jsonutils"
+	"github.com/asciich/asciichgolangpublic/pkg/fileformats/yamlutils"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 )
