+++ pkg/kubernetesutils/kubernetesimplementationindependend/ResourcesYaml.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/asciich/asciichgolangpublic/datatypes/stringsutils"
-	"github.com/asciich/asciichgolangpublic/fileformats/yamlutils"
+	"github.com/asciich/asciichgolangpublic/pkg/fileformats/yamlutils"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 	"gopkg.in/yaml.v3"
 )
